--- mobile/assets/grade-grade_fill1_symbol.svg:Zone.Identifier
@@ -1,3 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-HostUrl=about:internet
