--- mobile/assets/grade_FILL1_wght400_GRAD0_opsz24.svg:Zone.Identifier
@@ -1,3 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-HostUrl=about:internet
